requirements.txt
evdev==1.6.1
PyYAML==6.0
